refactor: move safe_get to utils package
