+++ ado_asana_sync/utils/utils.py
@@ -1,0 +1,19 @@
+def safe_get(obj, *attrs_keys):
+    """
+    Safely retrieves nested attributes from an object.
+
+    Args:
+        obj: The object to retrieve attributes from.
+        *attrs_keys: Variable number of attribute keys.
+
+    Returns:
+        The value of the nested attribute if found, else None.
+    """
+    for attr_key in attrs_keys:
+        if isinstance(obj, dict):
+            obj = obj.get(attr_key)
+        else:
+            obj = getattr(obj, attr_key, None)
+        if obj is None:
+            return None
+    return obj
